Cleaned up the request type
