+++ runtime/src/common.ts
@@ -2,3 +2,5 @@
  * This file exists to avoid errors on references to @sapper/common from inside Sapper.
  * (@sapper/common is a namespace defined in index.d.ts)
  */
+
+ export default null;
